# packages/shared/openapi.yaml
openapi: 3.0.3
info:
  title: Lony API
  version: 0.1.0
  description: Shared peer loan ledger. Phase 8 adds rate limits, idempotency, balance reconcile reporting, and legal disclaimer acceptance.
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /auth/verify:
    post:
      summary: Verify email ownership
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /auth/refresh:
    post:
      summary: Rotate refresh token and issue a new access token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
  /me:
    get:
      summary: Current profile
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Missing or invalid access token
    patch:
      summary: Update profile
      operationId: patchMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchMeRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required: [id, email, display_name, timezone, locale, email_verified, status, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        timezone: { type: string }
        locale: { type: string }
        default_currency_code: { type: string, nullable: true }
        email_verified: { type: boolean }
        status: { type: string }
        created_at: { type: string, format: date-time }
    UserResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/User" }
...
    VerifyRequest:
      type: object
      required: [email, code]
      properties:
        email: { type: string, format: email }
        code: { type: string, minLength: 6, maxLength: 6 }
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    TokenResponse:
      type: object
      required: [access_token, refresh_token, expires_in, token_type, user]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer }
        token_type: { type: string, example: Bearer }
        user: { $ref: "#/components/schemas/User" }
    PatchMeRequest:
      type: object
      properties:
        display_name: { type: string }
        timezone: { type: string }
        locale: { type: string }
        default_currency_code: { type: string }
...